backend/requirements.txt
fastapi>=0.95.0
uvicorn>=0.21.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.5
bcrypt>=4.0.1
motor>=3.1.1
pymongo>=4.3.3
pycryptodome>=3.10.1
python-dotenv>=0.19.0